requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.11.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
orjson==3.4.6
promise==2.3
protobuf==3.14.0
pydantic==1.7.3
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.1
Rx==1.6.1
six==1.15.0
starlette==0.13.6
ujson==3.2.0
urllib3==1.26.2
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
